+++ platforms/core-configuration/configuration-cache/src/integTest/groovy/org/gradle/internal/cc/impl/ConfigurationCacheIntegrationTest.groovy
@@ -24,51 +24,6 @@ import spock.lang.Issue
 
 class ConfigurationCacheIntegrationTest extends AbstractConfigurationCacheIntegrationTest {
 
-    def "humble beginnings"() {
-        given:
-        def configurationCache = newConfigurationCacheFixture()
-        file("gradle.properties") << "org.gradle.configuration-cache.entries-per-key=2"
-
-        and:
-        settingsFile ''
-
-        when:
-        configurationCacheRun 'help'
-
-        then:
-        configurationCache.assertStateStored()
-
-        when:
-        settingsFile '// a change'
-
-        and:
-        configurationCacheRun 'help'
-
-        then:
-        configurationCache.assertStateStored()
-
-        when: 'switching back to original settings file'
-        settingsFile.text = ''
-
-        and:
-        configurationCacheRun 'help'
-
-        then:
-        configurationCache.assertStateLoaded()
-
-        when: 'switching back to 2nd settings file'
-        settingsFile.text = '// a change'
-
-        and:
-        configurationCacheRun 'help'
-
-        then:
-        configurationCache.assertStateLoaded()
-        file('.gradle/configuration-cache').traverse {
-            println testDirectory.relativePath(it)
-        }
-    }
-
     def "configuration cache is out of incubation"() {
         given:
         settingsFile << ""
